+++ hdt-lib/src/libdcs/CSD_PFC.h
@@ -64,12 +64,12 @@ class CSD_PFC : public CSD
 	@s: the string to be located.
 	@len: the length (in characters) of the string s.
     */
-    uint32_t locate(const unsigned char *s, uint32_t len);
+    size_t locate(const unsigned char *s, size_t len);
 
     /** Returns the string identified by id.
 	@id: the identifier to be extracted.
     */
-    unsigned char * extract(uint32_t id);
+    unsigned char * extract(size_t id);
 
     void freeString(const unsigned char *str);
 
@@ -78,7 +78,7 @@ class CSD_PFC : public CSD
 	@dict: the plain uncompressed dictionary.
 	@return: number of total symbols in the dictionary.
     */
-    unsigned int decompress(unsigned char **dict);
+    size_t decompress(unsigned char **dict);
 
     /** Returns the size of the structure in bytes. */
     uint64_t getSize();
@@ -105,7 +105,7 @@ class CSD_PFC : public CSD
 
     uint32_t blocksize;	//! Number of strings stored in each block.
     hdt::LogSequence2 *blocks;	//! Start positions of each block in the encoded sequence.
-    uint32_t nblocks;   //! Number of blocks
+    size_t nblocks;   //! Number of blocks
 
     /** Locates the block in where the string 's' can be stored. This method is
 	based on a binary search comparing the first string in each block and
@@ -124,14 +124,14 @@ class CSD_PFC : public CSD
 	@len: the length (in characters) of the string s.
 	@return: the ID for 's' or 0 if it is not exist.
     */
-    unsigned int locateInBlock(size_t block, const unsigned char *s, unsigned int len);
+    size_t locateInBlock(size_t block, const unsigned char *s, size_t len);
 
     /** Extracts the o-th string in the given 'block'.
 	@block: block to be accesed.
 	@o: internal offset for the required string in the block.
 	@return: the extracted string.
     */
-    unsigned char *extractInBlock(unsigned int block, unsigned int o);
+    unsigned char *extractInBlock(size_t block, size_t o);
 
 
     /** Obtains the length of the long common prefix (lcp) of str1 and str2.
