+++ hdt-lib/src/triples/BitmapTriples.cpp
@@ -126,10 +126,10 @@ void BitmapTriples::load(ModifiableTriples &triples, ProgressListener *listener)
 	LogSequence2 *vectorY = new LogSequence2(bits(triples.getNumberOfElements()));
 	LogSequence2 *vectorZ = new LogSequence2(bits(triples.getNumberOfElements()),triples.getNumberOfElements());
 
-	unsigned int lastX, lastY, lastZ;
-	unsigned int x, y, z;
+    size_t lastX, lastY, lastZ;
+    size_t x, y, z;
 
-	unsigned int numTriples=0;
+    size_t numTriples=0;
 
 	while(it->hasNext()) {
 		TripleID *triple = it->next();
